AUTO : Forward from beta to master

+++ sample/mouse_position_translation/src/main.rs
@@ -23,7 +23,7 @@ fn main()
     height : DISPLAY_HEIGHT,
     resizable : false,
     .. Default::default()
-  })
+  } )
   .add_plugins( DefaultPlugins )
   .add_startup_system( setup )
   .add_system_set_to_stage
@@ -46,13 +46,13 @@ fn setup( mut commands : Commands )
 
 //
 
-fn cursor_system( window : Res< Windows >, q_camera : Query< &Transform, With<MainCamera > > )
+fn cursor_system( window : Res< Windows >, q_camera : Query< &Transform, With< MainCamera > > )
 {
   let primary_window = window.get_primary().unwrap();
 
   if let Some( pos ) = primary_window.cursor_position()
   {
-    let size = Vec2::new( primary_window.width() as f32, primary_window.height() as f32 );
+    let size = Vec2::new( primary_window.width(), primary_window.height() );
 
     // mouse coordinates related to center
     let p = pos - size / 2.0;
